Refactor preview code into service
